+++ src/rstgen_bypass.sv
@@ -47,8 +47,11 @@ module rstgen_bypass #(
             synch_regs_q <= {synch_regs_q[NumRegs-2:0], 1'b1};
         end
     end
-
+    // pragma translate_off
+    `ifndef VERILATOR
     initial begin : p_assertions
         if (NumRegs < 1) $fatal(1, "At least one register is required.");
     end
+    `endif
+    // pragma translate_on
 endmodule