+++ test/simulate.sh
@@ -35,3 +35,11 @@ done
 for num in 1 4 7; do
   call_vsim rr_arb_tree_tb -GNumInp=$num -coverage -voptargs="+acc +cover=bcesfx"
 done
+
+for spill_reg in 0 1; do
+  for num_inp in 1 4 18; do
+    for num_out in 1 4 18; do
+      call_vsim stream_xbar_tb -GNumInp=$num_inp -GNumOut=$num_out -GSpillReg=$spill_reg -coverage -voptargs="+acc +cover=bcesfx"
+    done
+  done
+done