+++ src/stream_demux.sv
@@ -8,24 +8,23 @@
 // CONDITIONS OF ANY KIND, either express or implied. See the License for the
 // specific language governing permissions and limitations under the License.
 
-/// Stream demultiplexer: Connects the input stream (valid-ready) handshake to one of `N_OUP` output
-/// stream handshakes.
+/// Connects the input stream (valid-ready) handshake to one of `N_OUP` output stream handshakes.
 ///
 /// This module has no data ports because stream data does not need to be demultiplexed: the data of
 /// the input stream can just be applied at all output streams.
-
 module stream_demux #(
-  parameter integer N_OUP = 1,
+  /// Number of connected outputs.
+  parameter int unsigned N_OUP     = 32'd1,
   /// Dependent parameters, DO NOT OVERRIDE!
-  parameter integer LOG_N_OUP = $clog2(N_OUP)
+  parameter int unsigned LOG_N_OUP = (N_OUP > 32'd1) ? unsigned'($clog2(N_OUP)) : 1'b1
 ) (
-  input  logic                  inp_valid_i,
-  output logic                  inp_ready_o,
+  input  logic                 inp_valid_i,
+  output logic                 inp_ready_o,
 
-  input  logic  [LOG_N_OUP-1:0] oup_sel_i,
+  input  logic [LOG_N_OUP-1:0] oup_sel_i,
 
-  output logic  [N_OUP-1:0]     oup_valid_o,
-  input  logic  [N_OUP-1:0]     oup_ready_i
+  output logic [N_OUP-1:0]     oup_valid_o,
+  input  logic [N_OUP-1:0]     oup_ready_i
 );
 
   always_comb begin