+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - stream_to_mem: Allows to use memories with flow control (req/gnt) for requests but
   without flow control for output data to be used in streams.
+- isochronous_spill_register: Isochronous clock domain crossing cutting all paths.
 
 ### Fixed
 - Improve tool compatibility.