+++ CHANGELOG.md
@@ -9,9 +9,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - stream_to_mem: Allows to use memories with flow control (req/gnt) for requests but
   without flow control for output data to be used in streams.
 - isochronous_spill_register: Isochronous clock domain crossing cutting all paths.
+- `rr_arb_tree_tb`: Systemverilog testbench for `rr_arb_tree`, which checks for fair throughput.
 
 ### Fixed
 - Improve tool compatibility.
+- `rr_arb_tree`: Add parameter `FairArb` to distribute throughput of input requests evenly when
+  not all inputs have requests active.
 
 ## 1.18.0 - 2020-04-15
 ### Added