+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `rr_arb_tree`: Properly degenerate `rr_i` and `idx_o` signals.
 - `rr_arb_tree`: Add parameter `FairArb` to distribute throughput of input requests evenly when
   not all inputs have requests active.
+- `stream_demux`: Properly degenerate `inp_sel_i` signal.
 
 ## 1.18.0 - 2020-04-15
 ### Added